Add interactive proof rounds

+++ humidor/src/f2_8x3_9.rs
@@ -2,23 +2,23 @@
 use proptest::{*, prelude::*};
 
 #[derive(Debug, Clone, Copy, Default)]
-pub struct F(pub i64);
+pub struct F(pub i128);
 
 impl F {
     pub const ZERO: Self = Self(0);
     pub const ONE: Self = Self(1);
-    pub const MOD: i64 = 5038849; // 2**8 * 3**9 + 1
-    pub const PHI: i64 = Self::MOD - 1;
+    pub const MOD: i128 = 5038849; // 2**8 * 3**9 + 1
+    pub const PHI: i128 = Self::MOD - 1;
     pub const GEN: Self = Self(29);
-    pub const PHI_2_EXP: i64 = 8;
-    pub const PHI_3_EXP: i64 = 9;
+    pub const PHI_2_EXP: i128 = 8;
+    pub const PHI_3_EXP: i128 = 9;
     pub const BITS: usize = 22; // floor(log2(MOD))
 
     // [ GEN**(PHI / (2**p)) % MOD | p <- [0 .. PHI_2_EXP] ]
-    pub const ROOTS_BASE_2 : [i64; Self::PHI_2_EXP as usize + 1] =
+    pub const ROOTS_BASE_2 : [i128; Self::PHI_2_EXP as usize + 1] =
         [1, 5038848, 4695727, 2653925, 4169754, 3835586, 461362, 1759713, 4318906];
     // [ GEN**(PHI / (2**p)) % MOD | p <- [0 .. PHI_3_EXP] ]
-    pub const ROOTS_BASE_3 : [i64; Self::PHI_3_EXP as usize + 1] =
+    pub const ROOTS_BASE_3 : [i128; Self::PHI_3_EXP as usize + 1] =
         [1, 2517480, 826430, 2522883, 1582093, 2737326, 1556385, 14459, 855021, 1814687];
 
     pub fn bytes(self) -> Vec<u8> { self.0.to_be_bytes().iter().cloned().collect() }
@@ -28,7 +28,7 @@ impl F {
         (Self::MOD + x).into()
     }
 
-    pub fn pow(self, mut e: i64) -> Self {
+    pub fn pow(self, mut e: i128) -> Self {
         debug_assert!(e >= 0); // Ensure logical shift
         let Self(mut b) = self;
         let mut acc = 1;
@@ -54,12 +54,12 @@ impl F {
     }
 }
 
-impl std::convert::From<i64> for F {
-    fn from (n: i64) -> F { F(n % Self::MOD) }
+impl std::convert::From<i128> for F {
+    fn from (n: i128) -> F { F(n % Self::MOD) }
 }
 
-impl std::convert::From<F> for i64 {
-    fn from (F(n): F) -> i64 { n }
+impl std::convert::From<F> for i128 {
+    fn from (F(n): F) -> i128 { n }
 }
 
 impl std::ops::Add for F {
@@ -123,7 +123,7 @@ impl num_traits::identities::Zero for F {
 impl num_traits::Num for F {
     type FromStrRadixErr = core::num::ParseIntError;
     fn from_str_radix(s: &str, r: u32) -> Result<Self, Self::FromStrRadixErr> {
-        <i64 as num_traits::Num>::from_str_radix(s, r).map(From::from)
+        <i128 as num_traits::Num>::from_str_radix(s, r).map(From::from)
     }
 }
 
@@ -138,7 +138,7 @@ impl Arbitrary for F {
     type Parameters = ();
     type Strategy = BoxedStrategy<Self>;
     fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
-        any::<i64>().prop_map(F::from).boxed()
+        any::<i128>().prop_map(F::from).boxed()
     }
 }
 
@@ -146,21 +146,21 @@ impl Arbitrary for F {
 proptest! {
     #[test]
     #[allow(non_snake_case)]
-    fn test_eq(x: i64) {
+    fn test_eq(x: i128) {
         let f = F::from(x);
         assert_eq!(f, f.positivize());
     }
 
     #[test]
     #[allow(non_snake_case)]
-    fn test_add_neg(x: i64) {
+    fn test_add_neg(x: i128) {
         let f = F::from(x);
         assert_eq!(f + f.neg(), F::ZERO);
     }
 
     #[test]
     #[allow(non_snake_case)]
-    fn test_mul_recip(x: i64) {
+    fn test_mul_recip(x: i128) {
         let f = F::from(x);
         assert_eq!(f * f.recip(), F::ONE);
     }
